extract out core health check service logic from osgi wrapper
